all changes made after review
