+++ .gitignore
@@ -1,0 +1,19 @@
+.idea/
+*.sublime-*
+*.DS_Store
+.project
+.pydevproject
+.settings/
+.Python
+.cache/
+local.ini
+
+*.egg
+*.egg-info
+*.eggs
+*.pyc
+*.swp
+*.swo
+*.exe
+
+.testcases/pytest_cache/
